reset

+++ database/migrations/2023_08_21_201550_create_product__reviews_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('user_id');
             $table->text('review');
-            $table->integer('rating'); // Scale from 1 to 5
+            $table->decimal('rating'); // Scale from 1 to 5
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade')->onUpdate('cascade');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
             $table->timestamps();
